user login and last login time update on database

+++ src/components/Users.jsx
@@ -1,23 +1,40 @@
 import { data } from "autoprefixer";
 import React, { useState } from "react";
 import { useLoaderData } from "react-router-dom";
+import Swal from "sweetalert2";
 
 const Users = () => {
   const loadedUsers = useLoaderData();
   const [users, setUsers] = useState(loadedUsers);
   const handleDeleteUser = (id) => {
-    fetch(`http://localhost:5000/user/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        console.log(data);
-        if (data.deletedCount > 0) {
-          alert("user delete Successfully");
-          const remaining = users.filter((user) => user._id !== id);
-          setUsers(remaining);
-        }
-      });
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:5000/user/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            console.log(data);
+            if (data.deletedCount > 0) {
+              Swal.fire({
+                title: "Deleted!",
+                text: "Your file has been deleted.",
+                icon: "success",
+              });
+            }
+          });
+        const remaining = users.filter((user) => user._id !== id);
+        setUsers(remaining);
+      }
+    });
   };
   return (
     <div>
@@ -44,7 +61,7 @@ const Users = () => {
                 <td>{user._id}</td>
                 <td>{user.email}</td>
                 <td>{user.creationTime}</td>
-                <td>Blue</td>
+                <td>{user.lastSignInTime}</td>
                 <td onClick={() => handleDeleteUser(user._id)} className='btn '>
                   X
                 </td>
